refactor(render): rename shared writer to relational_io; preserve to_dataframes history (self-review) [LTV-Pf.2]
Self-review caught that the relational.py split lost git history and left two
files named relational.py. Both files shared the basename, so git's same-path
heuristic recorded the scheme file (which inherits to_dataframes + _TABLE_SOURCES,
the bulk of the original) as a brand-new ADD and the gutted shared remnant as a
MODIFY — breaking `git blame`/`log --follow` on to_dataframes, and contradicting
the history-preservation standard Pf.1 set.

Fix: rename the shared writer module leadforge/render/relational.py →
leadforge/render/relational_io.py. With the basename collision gone, git now
records the scheme file as R067 (rename of the original — to_dataframes history
preserved back to M8 #13) and relational_io.py as a clean ADD. Bonus: no more
ambiguous duplicate `relational.py` (the shared writer is scheme-agnostic and
isn't "relational"-specific anyway).

- Importers of write_relational_tables updated (scheme __init__, test).
- test_module_layout.py: assert the shared writer is at render.relational_io and
  the flat `leadforge.render.relational` module is gone.
- CHANGELOG "Moved" table + CLAUDE.md (both layout sections) updated.

Verified byte-identical (14/14); full suite 1532 passed / 51 skipped; ruff +
mypy clean.
