+++ CHANGELOG.md
@@ -25,10 +25,14 @@ back-compat shims, by design):
 | `leadforge.render.relational_snapshot_safe` | `leadforge.schemes.lead_scoring.render.relational_snapshot_safe` |
 | `leadforge.render.tasks` | `leadforge.schemes.lead_scoring.render.tasks` |
 | `leadforge.render.relational:to_dataframes` | `leadforge.schemes.lead_scoring.render.relational:to_dataframes` |
+| `leadforge.render.relational:write_relational_tables` | `leadforge.render.relational_io:write_relational_tables` |
 
-`leadforge.render` remains the shared bundle-output envelope:
-`leadforge.render.relational.write_relational_tables` (the scheme-agnostic table
-writer) and `leadforge.render.manifests` stay put.  The lead-scoring `schema`
+The flat `leadforge.render.relational` module is **removed**: its 9-table
+assembler (`to_dataframes`) moved to the scheme, and the scheme-agnostic writer
+(`write_relational_tables`) moved to the new `leadforge.render.relational_io`
+(renamed to avoid a basename clash with the scheme's `relational.py`).
+`leadforge.render` remains the shared bundle-output envelope
+(`relational_io` + `manifests`).  The lead-scoring `schema`
 specs relocate in a follow-up PR (LTV-Pg).  Consumers importing internals (e.g.
 the `leadforge-datasets-private` build scripts) must update to the new paths;
 the package stays on the `1.x` line (the public contract did not change).
